requirements.txt
langchain-google-genai
browser_use
python-dotenv
pydantic
flask